Improving AdditiveSceneLoader usability
* Renamed class and file to reflect naming conventions
* Added namespace
* Replaced magic string and const with serialized fields
